requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
pydantic==1.10.12
python-dotenv==1.0.0
sqlalchemy==2.0.22
mysql-connector-python==8.1.1
langchain==0.1.0
langchain-community==0.1.0
langchain-google-genai==0.1.0
langgraph==0.1.0
